Fixed some game flow issues. Game end is ok. Some kind of bug in maxi yatzy.
