+++ src/enjug/erijan/games/yatzy/model/GameStateObserver.java
@@ -4,7 +4,7 @@ package enjug.erijan.games.yatzy.model;
  * Created by Jan Eriksson on 30/10/15.
  */
 
-public interface StateInfoObserver {
-  public void update(StateInfo stateInfo);
+public interface GameStateObserver {
+  public void update(GameState gameState);
 }
