--- src/enjug/erijan/games/yatzy/model/StateInfoSubject.java
@@ -1,11 +1,0 @@
-package enjug.erijan.games.yatzy.model;
-
-/**
- * Created by Jan Eriksson on 30/10/15.
- */
-
-public interface StateInfoSubject {
-  public void registerObserver(StateInfoObserver o);
-  public void removeObserver(StateInfoObserver o);
-  public void notifyObservers();
-}
